update
git-svn-id: http://svn.tudos.org/repos/tudos/trunk@240 a704ac0b-3a55-4d43-a2a9-7be6f07c34fb

+++ l4/pkg/pingpong/server/src/global.c
@@ -4,7 +4,7 @@
 #include "global.h"
 
 void*
-memchr(const void *s, int c, unsigned n) 
+memchr(const void *s, int c, size_t n) 
 {
   register const char* t=s;
   int i;
